requirements.txt
astroid==2.4.2
autopep8==1.5.4
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
colorama==0.4.4
cryptography==3.4.8
distlib==0.3.2
filelock==3.0.12
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.26
pdfdocument==4.0.0
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.3.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
reportlab==3.6.1
rope==0.18.0
six==1.15.0
toml==0.10.1
virtualenv==20.7.2
virtualenv-clone==0.5.6
wrapt==1.12.1
python_decouple==3.4

requests~=2.26.0
numpy~=1.21.2


